DP-Andres/requirements.txt
Flask==3.0.3
google_api_python_client==2.129.0
google_auth_oauthlib==1.2.0
gspread==6.1.2
numpy==1.26.4
qrcode==7.4.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
pandas==2.2.2
Pillow==10.4.0
backoff==2.2.1
urllib3==1.26.15
requests==2.28.2
Requests==2.32.3
streamlit==1.38.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
reportlab ==4.2.2
bs4
pyzbar
plotly
psutil
